% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_arcgis_deployments.R
\name{cb_arcgis_deployments}
\alias{cb_arcgis_deployments}
\title{Pull ARU deployment information entered into ArcGIS Field Maps app}
\usage{
cb_arcgis_deployments(object)
}
\arguments{
\item{object}{}
}
\description{
Pull ARU deployment information entered into ArcGIS Field Maps app
}
