% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_aru_lat_long.R
\name{cb_aru_lat_long}
\alias{cb_aru_lat_long}
\title{Extract lat/long coordinates from an ARU sf point object}
\usage{
cb_aru_lat_long(arus)
}
\arguments{
\item{arus}{A sf point object of ARU deployments in WGS 84 coordinate system}
}
\description{
Extract lat/long coordinates from an ARU sf point object
}
