% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_build_s3_db.R
\name{cb_build_s3_db}
\alias{cb_build_s3_db}
\title{Create a database of UW-Madison S3 contents that is easy to query}
\usage{
cb_build_s3_db(s3_df, db_path)
}
\arguments{
\item{s3_df}{\code{Tibble} of S3 contents from \code{cb_get_s3_df}}

\item{db_path}{Location for storing \code{DuckDB}}
}
\value{
\code{DuckDB} containing S3 file paths in a table called \code{s3_contents}
}
\description{
Create a database of UW-Madison S3 contents that is easy to query
}
\examples{
\dontrun{
# build database
cb_build_s3_db(s3_df = s3_df, db_path = "Z:/Bioacoustics/s3_bucket_contents")
}
}
