% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_cells_by_ownership.R
\name{cb_cells_by_ownership}
\alias{cb_cells_by_ownership}
\title{Get a list of cell IDs depending on land ownership}
\usage{
cb_cells_by_ownership(ownership)
}
\arguments{
\item{ownership}{US Forest Service (\code{usfs}), National Park Service (\code{nps}), or all cells (\code{any})}
}
\value{
List of cell IDs
}
\description{
Get a list of cell IDs depending on land ownership
}
\examples{
\dontrun{
cell_list <- cb_cells_by_ownership('usfs')
print(cell_list)
}
}
