% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_check_empty_folders.R
\name{cb_check_empty_folders}
\alias{cb_check_empty_folders}
\title{Check for folders missing acoustic data on research drive}
\usage{
cb_check_empty_folders(root)
}
\arguments{
\item{root}{Diretory of FLAC or JSON files}
}
\description{
Check for folders missing acoustic data on research drive
}
\examples{
\dontrun{
# path to FLACs or JSONs
root <- "Z:/Acoustic_Data/ARU_Data_Raw/ARU_Sierra_Monitoring/2025"

# check for empty folders (usually write this to a csv)
cb_check_empty_folders(root)
}
}
