% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_check_s3.R
\name{cb_check_s3}
\alias{cb_check_s3}
\title{Check that Amazon S3 bucket exists}
\usage{
cb_check_s3()
}
\value{
Output will contain \code{TRUE} if connection to S3 bucket is working
}
\description{
Check that Amazon S3 bucket exists
}
\examples{
\dontrun{
cb_check_s3()
}
}
