% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_check_secondary_sampling.R
\name{cb_check_secondary_sampling}
\alias{cb_check_secondary_sampling}
\title{Title}
\usage{
cb_check_secondary_sampling(start_date, end_date, occasion_length)
}
\arguments{
\item{occasion_length}{}
}
\description{
Title
}
