% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_clear_hard_drive.R
\name{cb_clear_hard_drive}
\alias{cb_clear_hard_drive}
\title{Clear contents of an external hard drive}
\usage{
cb_clear_hard_drive(path)
}
\arguments{
\item{path}{}
}
\description{
Clear contents of an external hard drive
}
