% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_clear_hd.R
\name{cb_clear_hd}
\alias{cb_clear_hd}
\title{Clear contents of an external hard drive}
\usage{
cb_clear_hd(path)
}
\arguments{
\item{path}{External hard drive directory}
}
\value{
Empty external hard drive
}
\description{
Clear contents of an external hard drive
}
\examples{
\dontrun{
cb_clear_hd('D:/')
}
}
