% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_config_occ_project.R
\name{cb_config_occ_project}
\alias{cb_config_occ_project}
\title{Set parameters for an occupancy analysis}
\usage{
cb_config_occ_project(
  start_year,
  end_year,
  species,
  template_used,
  study_type,
  cell_list
)
}
\arguments{
\item{start_year}{Start year}

\item{end_year}{End year}

\item{species}{Species (CSOW or BDOW)}

\item{template_used}{BirdNET version used}

\item{study_type}{Study type (e.g., Sierra Monitoring, Coastal Barred)}

\item{cell_list}{List of cell IDs to query (C0001, C0002, ...)}
}
\description{
Set parameters for an occupancy analysis
}
