% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_pam_config.R
\name{cb_config_pam_project}
\alias{cb_config_pam_project}
\title{Write configuration parameters for PAM R project}
\usage{
cb_config_pam_project(study_year, study_type, study_path)
}
\arguments{
\item{study_path}{}
}
\description{
Write configuration parameters for PAM R project
}
\examples{
\dontrun{
cb_config_pam_project(
  study_year = lubridate::year(Sys.Date()),
  study_type = 'Sierra_Monitoring',
  study_path = 'Z:'
)
}
}
