% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_connect_db.R
\name{cb_connect_db}
\alias{cb_connect_db}
\title{Connect to the UW spotted owl database}
\usage{
cb_connect_db()
}
\value{
\code{DBI} database connection
}
\description{
Connect to the UW spotted owl database
}
\examples{
\dontrun{
# connect to the database
cb_connect_db()
}
}
