% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_pam_project.R
\name{cb_create_pam_project}
\alias{cb_create_pam_project}
\title{Create R project directory for processing PAM program data}
\usage{
cb_create_pam_project(project_name, base_dir = getwd())
}
\value{
R project
}
\description{
Create R project directory for processing PAM program data
}
\examples{
\dontrun{
cb_create_pam_project(
  project_name = "2026_sierra_acoustic_data_processing",
  base_dir = "Z:/"
)
}
}
