% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_selection_directories.R
\name{cb_create_selection_directories}
\alias{cb_create_selection_directories}
\title{Create directories to store selection tables}
\usage{
cb_create_selection_directories(output, directory, species_group)
}
\arguments{
\item{output}{Main directory to store selection tables}

\item{directory}{Directories to store selection tables by group and cell (e.g., G001/C0001)}

\item{species_group}{Make selection tables for CSOW/BDOW or forest owls}
}
\value{
Folders to hold selection tables
}
\description{
Create directories to store selection tables
}
