% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_disconnect_s3_db.R
\name{cb_disconnect_s3_db}
\alias{cb_disconnect_s3_db}
\title{Disconnect from S3 database}
\usage{
cb_disconnect_s3_db()
}
\description{
Disconnect from S3 database
}
\examples{
\dontrun{
# disconnect from S3 database
cb_disconnect_s3_db()
}
}
