% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filename_to_datetime.R
\name{cb_filename_to_datetime}
\alias{cb_filename_to_datetime}
\title{Extract date/datetime from a path or filename}
\usage{
cb_filename_to_datetime(x, type = c("datetime", "date"))
}
\arguments{
\item{x}{Path or filename}

\item{type}{Datetime or date}
}
\value{
Datetime or date
}
\description{
Extract date/datetime from a path or filename
}
