% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filename_to_datetime_duckdb.R
\name{cb_filename_to_datetime_duckdb}
\alias{cb_filename_to_datetime_duckdb}
\title{Extract date/datetime from a path or filename (\code{duckdb} translation)}
\usage{
cb_filename_to_datetime_duckdb(x, type = c("datetime", "date"))
}
\arguments{
\item{type}{}
}
\description{
Extract date/datetime from a path or filename (\code{duckdb} translation)
}
