% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_flac_duration_s3.R
\name{cb_flac_duration_s3}
\alias{cb_flac_duration_s3}
\title{Get duration of FLAC files (in seconds) stored on S3}
\usage{
cb_flac_duration_s3(s3, object, metadata = FALSE)
}
\arguments{
\item{s3}{S3 endpoint}

\item{object}{S3 key}

\item{metadata}{If FALSE, return just FLAC file duration}
}
\description{
Get duration of FLAC files (in seconds) stored on S3
}
