% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_audio_duration.R
\name{cb_get_audio_duration}
\alias{cb_get_audio_duration}
\title{Get duration of audio file (in minutes)}
\usage{
cb_get_audio_duration(path)
}
\arguments{
\item{path}{Path to audio file}
}
\value{
Duration of audio file in minutes
}
\description{
Get duration of audio file (in minutes)
}
\examples{
\dontrun{
flac_duration_df <-
  flac_df |>
    # list of flac paths
    pull(path) |>
    # run in parallel
    future_map_dfr(get_duration)
    }
}
