% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_coords.R
\name{cb_get_coords}
\alias{cb_get_coords}
\title{Get x/y coordinates added as columns to an \code{sf} object}
\usage{
cb_get_coords(df)
}
\arguments{
\item{df}{\code{sf} POINT object}
}
\value{
\code{sf} POINT object with columns for x/y coordinate values
}
\description{
Get x/y coordinates added as columns to an \code{sf} object
}
\examples{
\dontrun{
# create db connection
cb_connect_db()

# read in 2024 sierra monitoring deployments
deployments_df <-
  tbl(conn, "acoustic_field_visits") |>
  filter(survey_year == 2024 & study_type == 'Sierra Monitoring') |>
  select(
    id,
    study_type,
    group_id,
    visit_id,
    cell_id,
    unit_number,
    deployment_name,
    survey_year,
    deploy_date,
    recover_date,
    utm_zone,
    utme,
    utmn
  ) |>
  # pull into memory
  collect()

# disconnect now
cb_disconnect_db()

# make ARUs sf, convert to WGS 84
deployments_sf <-
  deployments_df |>
  cb_make_aru_sf()

# add coordinates
deployments_sf <-
  deployments_sf |>
  cb_get_coords()
}
}
