% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_date_code_output.R
\name{cb_get_date_code_output}
\alias{cb_get_date_code_output}
\title{Get date of most recent code run}
\usage{
cb_get_date_code_output(folder)
}
\arguments{
\item{folder}{A code output folder (e.g., pre_birdnet, post_birdnet).}
}
\description{
Get date of most recent code run
}
