% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_date_time.R
\name{cb_get_date_time}
\alias{cb_get_date_time}
\title{Get datetime in character format for current or most recent code run}
\usage{
cb_get_date_time(type, folder)
}
\arguments{
\item{type}{Current datetime or datetime of most recent code run}

\item{folder}{Directory to find datetime if looking for most recent one}
}
\value{
Datetime in character format
}
\description{
Get datetime in character format for current or most recent code run
}
