% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_surveyor_overlap.R
\name{cb_get_surveyor_overlap}
\alias{cb_get_surveyor_overlap}
\title{Identify BirdNET selections that overlapped with CSOW broadcast surveys}
\usage{
cb_get_surveyor_overlap(
  df,
  human_hoots,
  year,
  previous_date_time,
  new_date_time
)
}
\arguments{
\item{df}{}
}
\description{
Identify BirdNET selections that overlapped with CSOW broadcast surveys
}
