% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_list_db_tables.R
\name{cb_list_db_tables}
\alias{cb_list_db_tables}
\title{List owl database tables}
\usage{
cb_list_db_tables()
}
\value{
Character vector of database table names
}
\description{
List owl database tables
}
\examples{
\dontrun{
cb_list_db_tables()
}
}
