% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_list_s3_files.R
\name{cb_list_s3_files}
\alias{cb_list_s3_files}
\title{Get a list of S3 keys using AWS CLI}
\usage{
cb_list_s3_files(bucket, url, output_file)
}
\arguments{
\item{bucket}{S3 bucket name}

\item{url}{S3 url}

\item{output_file}{Text file of S3 keys}
}
\value{
A \code{tibble} of S3 keys and metadata
}
\description{
Get a list of S3 keys using AWS CLI
}
\examples{
\dontrun{
bucket <- "mpeery-archive"
url <- "https://s3.drive.wisc.edu"
output_file <- "C:/Users/jwiniarski/Desktop/s3_files.txt"

tictoc::tic()
cb_list_s3_files(
  bucket = bucket,
  url = url,
  output_file = output_file
  )
tictoc::toc()
}
}
