% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_logistic_to_threshold.R
\name{cb_logistic_to_threshold}
\alias{cb_logistic_to_threshold}
\title{Calculate BirdNET logit threshold(s) from logistic model of manually validated detections and their BirdNET logit score}
\usage{
cb_logistic_to_threshold(intercept, beta, p_tp)
}
\arguments{
\item{intercept}{Intercept value from logistic regression model}

\item{beta}{Slope value from logistic regression model}

\item{p_tp}{Probability of true positive detection value(s)}
}
\value{
\code{Tibble} containing true positive detection probability(ies) and associated BirdNET logit threshold(s)
}
\description{
Calculate BirdNET logit threshold(s) from logistic model of manually validated detections and their BirdNET logit score
}
\examples{
\dontrun{
cb_logistic_to_threshold(
  intercept = 0.830385299,
  beta = 1.111106172,
  p_tp = c(0.90, 0.95, 0.975, 0.99)
  )
}
}
