% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_make_aru_sf.R
\name{cb_make_aru_sf}
\alias{cb_make_aru_sf}
\title{Convert a data frame containing ARU UTM information into a simple feature object in WGS84}
\usage{
cb_make_aru_sf(df)
}
\arguments{
\item{df}{A data frame containing at minimum a column for the UTM zone (10 or 11; \code{utm_zone}) and easting (\code{utme}) and northing (\code{utmn}) coordinates}
}
\value{
A simple feature point object
}
\description{
Convert a data frame containing ARU UTM information into a simple feature object in WGS84
}
\examples{
\dontrun{
# example ARU data frame
aru_df <-
tibble::tribble(
  ~cell_id, ~utm_zone, ~utme, ~utmn,
  'C0001', 10, 670282, 4390557,
  'C0001', 10, 670972, 4390553,
  'C0001', 10, 669829, 4390388,
  'C9999', 11, 377972, 3919578
)
# convert to simple feature point geometry
arus_sf <-
  aru_df |>
    cb_make_aru_sf

# map
mapview::mapview(arus_sf)
}
}
