% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_occ_start_end_dates.R
\name{cb_occ_start_end_dates}
\alias{cb_occ_start_end_dates}
\title{Title}
\usage{
cb_occ_start_end_dates(aru_dates_df, occasion_length)
}
\arguments{
\item{occasion_length}{}
}
\description{
Title
}
