% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_play_audio.R
\name{cb_play_audio}
\alias{cb_play_audio}
\title{Play an audio file from R using SoX}
\usage{
cb_play_audio(audio_path, begin_time, end_time)
}
\arguments{
\item{audio_path}{Path to audio file}

\item{begin_time}{When to start playing audio}

\item{end_time}{When to stop playing audio file}
}
\value{
Audio at particular start/end time
}
\description{
Play an audio file from R using SoX
}
\examples{
\dontrun{
# download an audio file from xeno-canto
tmp_file <- tempfile(fileext = ".wav")

# store as a temporary file
download.file(
  "https://xeno-canto.org/877031/download",
  destfile = tmp_file,
  mode = ”wb”
)

# play the audio with SoX (this function can also play wavs)
cb_play_flac(
  flac_path = tmp_file,
  begin_time = 0,
  end_time = 10
)
}
}
