% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_query_effort_detections_occ.R
\name{cb_query_effort_detections_occ}
\alias{cb_query_effort_detections_occ}
\title{Title}
\usage{
cb_query_effort_detections_occ(
  species,
  template,
  study_type,
  cell_ids,
  start_year,
  end_year,
  start_date,
  end_date,
  occasion_length,
  num_occasions,
  type
)
}
\arguments{
\item{type}{}
}
\description{
Title
}
