% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_query_multi_state_occ.R
\name{cb_query_multi_state_occ}
\alias{cb_query_multi_state_occ}
\title{Generate encounter histories and effort data for multi-state, multi-season (dynamic) occupancy analysis}
\usage{
cb_query_multi_state_occ(
  species,
  template_used,
  study_type,
  cell_ids,
  start_year,
  end_year,
  start_date,
  end_date,
  occasion_length,
  num_occasions
)
}
\arguments{
\item{species}{Species (for now just CSOW)}

\item{template_used}{BirdNET version used}

\item{study_type}{Study type (e.g., Sierra Monitoring)}

\item{cell_ids}{Cells to query}

\item{start_year}{Start year}

\item{end_year}{End year}

\item{start_date}{Earliest date in the season to query}

\item{end_date}{Latest date in the season to query}

\item{occasion_length}{Length of secondary sampling occasion (in days)}

\item{num_occasions}{Number of secondary sampling occasions}
}
\value{
List of three data frames containing encounter histories (long and wide format) and effort (wide format)
}
\description{
Generate encounter histories and effort data for multi-state, multi-season (dynamic) occupancy analysis
}
\examples{
\dontrun{
start_year <- 2021
end_year <- 2025
study_type <- 'Sierra Monitoring'
species <- 'CSOW'
template_used <- 'SierraModel-1.0'

# get list of cell IDs (this will get all usfs cells)
cell_list <- cb_cells_by_ownership('usfs')

# determine start/end dates of primary time periods
aru_dates_df <-
  cb_season_dates(
    study_type = study,
    cell_ids = cell_list,
    start_year = start_year,
    end_year = end_year
    )

# set occasion length (in days)
occasion_length <- 7

# see if min/max dates work as is;
# otherwise this function will adjust the end date so there are N even length
# secondary sampling occasions
primary_season_dates_df <-
  cb_occ_start_end_dates(
    aru_dates_df = aru_dates_df,
    occasion_length = occasion_length
  )

# now set these for querying the database and formatting encounter histories
start_date <- primary_season_dates_df$start_date |> str_sub(5, 10)
end_date <- primary_season_dates_df$end_date |> str_sub(5, 10)
num_occasions <- primary_season_dates_df$num_occasions

# connect to database
cb_connect_db()

# now query
occ_df <-
  cb_query_multi_state_occ(
    species = species,
    template_used = template_used,
    study_type = study_type,
    cell_ids = cell_list,
    start_year = start_year,
    end_year = end_year,
    start_date = start_date,
    end_date = end_date,
    occasion_length = occasion_length,
    num_occasions = num_occasions
  )

# and disconnect
cb_disconnect_db()
}
}
