% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_read_birdnet_json_s3.R
\name{cb_read_birdnet_json_s3}
\alias{cb_read_birdnet_json_s3}
\title{Read a BirdNET JSON file (from S3) in tidy format}
\usage{
cb_read_birdnet_json_s3(object_key)
}
\arguments{
\item{object_key}{Path to BirdNET JSON (json.gz) on S3}
}
\value{
\code{Tibble} form of BirdNET JSON output
}
\description{
Read a BirdNET JSON file (from S3) in tidy format
}
\examples{
\dontrun{
# key on S3 to example JSON
json_path <-
  "Acoustic_Data/ARU_Data_Processed/BirdNET_Results_JSON/JSON_Sierra_Monitoring/2025/North/G010_V1/G010_V1_C0088_U2/G010_V1_C0088_U2_20250508/G010_V1_C0088_U2_20250508_060002Z.json.gz"

# read JSON
birdnet_df <- cb_read_birdnet_json_s3(json_path)
}
}
