% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_read_birdnet_selections.R
\name{cb_read_birdnet_selections}
\alias{cb_read_birdnet_selections}
\title{Read a BirdNET selection table}
\usage{
cb_read_birdnet_selections(birdnet_path)
}
\arguments{
\item{birdnet_path}{Path to a BirdNET selection table}
}
\value{
Tibble version of a BirdNET selection table
}
\description{
Read a BirdNET selection table
}
\examples{
\dontrun{
# set path to BirdNET selection file
birdnet_path <- here::here('G005_V1_C0125_U2_20230521_210001.BirdNET')

# read in BirdNET file
birdnet_selection_df <- cb_read_birdnet_selections(birdnet_path)
}
}
