% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_read_surveyor_data.R
\name{cb_read_surveyor_data}
\alias{cb_read_surveyor_data}
\title{Read broadcast survey data (for hoot removal)}
\usage{
cb_read_surveyor_data(path)
}
\arguments{
\item{path}{Path to surveyor data}
}
\value{
Tibble containing lat/long and survey date information
}
\description{
Read broadcast survey data (for hoot removal)
}
\examples{
\dontrun{
# path to data
path <- here::here('surveyor_data_example.csv')
surveyor_df <- cb_read_surveyor_data(path)
}
}
