% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_rename_files.R
\name{cb_rename_files}
\alias{cb_rename_files}
\title{Rename FLACs, JSONs, and BirdNET selection tables (and associated directories) with incorrect deployment information}
\usage{
cb_rename_files(dir, file_type, pattern, replacement)
}
\arguments{
\item{dir}{Directory where files are kept (for example, folder containing all FLACs for an ARU deployment)}

\item{file_type}{.flac, .json.gz, or .BirdNET}

\item{pattern}{String needing replacing}

\item{replacement}{String to replace it with}
}
\value{
Renamed FLAC files in renamed directories
}
\description{
Rename FLACs, JSONs, and BirdNET selection tables (and associated directories) with incorrect deployment information
}
\examples{
\dontrun{
# correct G001_C0001_V1_U2 to G001_C0001_V1_U1 for a set of FLACs
cb_rename_files(
  dir = 'path/to/deployments/G001_C0001_V1_U2',
  file_type = '.flac',
  pattern = 'U2',
  replacement = 'U1'
  )
}
}
