% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_rename_flacs.R
\name{cb_rename_flacs}
\alias{cb_rename_flacs}
\title{Rename flacs files with incorrect deployment information}
\usage{
cb_rename_flacs(dir, pattern, replacement)
}
\arguments{
\item{replacement}{}
}
\description{
Rename flacs files with incorrect deployment information
}
