% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_robocopy.R
\name{cb_robocopy}
\alias{cb_robocopy}
\title{Use \code{robocopy} to copy audio data from external hard drives}
\usage{
cb_robocopy(
  source,
  destination,
  threads = 16,
  retries = 3,
  wait = 5,
  log_file,
  show_output = TRUE
)
}
\arguments{
\item{show_output}{}
}
\description{
Use \code{robocopy} to copy audio data from external hard drives
}
