% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_s3_flac_duration.R
\name{cb_s3_flac_duration}
\alias{cb_s3_flac_duration}
\title{Get duration of FLAC files stored on S3}
\usage{
cb_s3_flac_duration(object)
}
\arguments{
\item{path}{}
}
\description{
Get duration of FLAC files stored on S3
}
