% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_sampling_summary.R
\name{cb_sampling_summary}
\alias{cb_sampling_summary}
\title{Get summary of sampling effort}
\usage{
cb_sampling_summary(
  species,
  template_used,
  study_type,
  cell_ids,
  start_year,
  end_year
)
}
\arguments{
\item{species}{Species to query (CSOW or BDOW)}

\item{template_used}{Template used}

\item{study_type}{Study type (e.g., Sierra Monitoring)}

\item{cell_ids}{List of cell IDs to query}

\item{start_year}{Start year to query}

\item{end_year}{End year to query}
}
\value{
\code{tibble} of sampling effort statistics
}
\description{
Get summary of sampling effort
}
