% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_save_s3_flac.R
\name{cb_save_s3_flac}
\alias{cb_save_s3_flac}
\title{Save FLAC files from UW-Madison S3 locally. If needing to download a large batch of S3 FLACs, best to use \code{cb_globus_transfer}}
\usage{
cb_save_s3_flac(file, destination_folder)
}
\arguments{
\item{file}{Path to FLAC file on  UW-Madison S3}

\item{destination_folder}{Path to folder where FLAC will be downloaded}
}
\value{
FLAC file
}
\description{
Save FLAC files from UW-Madison S3 locally. If needing to download a large batch of S3 FLACs, best to use \code{cb_globus_transfer}
}
\examples{
\dontrun{
# path to FLAC on S3
amazon_s3_key <- '/Acoustic_Data/ARU_Data/ARU_Sierra_Monitoring/2023/North/G002_V1/G002_V1_C0142_U2/G002_V1_C0142_U2_20230509/G002_V1_C0142_U2_20230509_000100Z.flac'
# download it
cb_save_s3_flac(amazon_s3_key, here::here())
}
}
