% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_save_s3_object.R
\name{cb_save_s3_object}
\alias{cb_save_s3_object}
\title{Save an object from S3 locally}
\usage{
cb_save_s3_object(s3_key, output_path)
}
\arguments{
\item{s3_key}{Key (i.e., file path) on S3}

\item{output_path}{Where to save the file locally}
}
\value{
S3 object
}
\description{
Save an object from S3 locally
}
\examples{
\dontrun{
# connect to S3 database
cb_connect_s3_db(db_path = "Z:/Bioacoustics/Acoustic_Code/s3_bucket_contents")

# pull a random S3 key
s3_key <-
  conn_s3 |>
    tbl('s3_contents') |>
    filter(regexp_matches(key, 'flac|json.gz')) |>
    slice_sample(n = 1) |>
    pull(key)

# save it locally with the same file name
cb_save_s3_object(s3_key = s3_key, output_path = here::here(basename(s3_key)))
}
}
