% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_segment_flac.R
\name{cb_segment_flac}
\alias{cb_segment_flac}
\title{Segment a FLAC file using the begin/end time of a BirdNET selection}
\usage{
cb_segment_flac(input_flac_path, output_flac_path, begin_time, end_time)
}
\arguments{
\item{input_flac_path}{Path for input FLAC file}

\item{output_flac_path}{Path to store FLAC segment}

\item{begin_time}{Time from start of recording (beginning of a selection)}

\item{end_time}{Time from start of recording (end of selection)}
}
\value{
FLAC file segment
}
\description{
Segment a FLAC file using the begin/end time of a BirdNET selection
}
