% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_set_db_credentials.R
\name{cb_set_db_credentials}
\alias{cb_set_db_credentials}
\title{Establish credentials for the UW spotted owl database}
\usage{
cb_set_db_credentials(credentials_path)
}
\value{
Saved passwords in keyring for spotted owl database
}
\description{
Establish credentials for the UW spotted owl database
}
\examples{
\dontrun{
# automatically set database credentials with csv
cb_set_db_credentials('read_only_credentials.csv')
}
}
