% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_set_future_cores.R
\name{cb_set_future_cores}
\alias{cb_set_future_cores}
\title{Set number of cores to run functions in parallel with future/furrr}
\usage{
cb_set_future_cores(n)
}
\arguments{
\item{n}{}
}
\description{
Set number of cores to run functions in parallel with future/furrr
}
