% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_set_s3_credentials.R
\name{cb_set_s3_credentials}
\alias{cb_set_s3_credentials}
\title{Establish credentials to access Amazon S3 bucket}
\usage{
cb_set_s3_credentials()
}
\description{
Establish credentials to access Amazon S3 bucket
}
\examples{
\dontrun{
# this will prompt to enter 1) the AWS access key ID, and 2) AWS secret access key
cb_set_s3_credentials()
}
}
