% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_show_globus_endpoints.R
\name{cb_show_globus_endpoints}
\alias{cb_show_globus_endpoints}
\title{Print Globus endpoint IDs}
\usage{
cb_show_globus_endpoints()
}
\value{
UUIDs for available Globus endpoints
}
\description{
Print Globus endpoint IDs
}
