% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_update_birdnet_db.R
\name{cb_update_birdnet_db}
\alias{cb_update_birdnet_db}
\title{Append new detections to a duckdb of all BirdNET species detections}
\usage{
cb_update_birdnet_db(spp_predictions_df, effort_df)
}
\value{
Updated duckdb
}
\description{
Append new detections to a duckdb of all BirdNET species detections
}
