% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_write_pipeline_batches.R
\name{cb_write_pipeline_batches}
\alias{cb_write_pipeline_batches}
\title{Write effort pipeline csv files}
\usage{
cb_write_pipeline_batches(df, output_path)
}
\arguments{
\item{output_path}{}
}
\value{
csv files of effort (i.e., row for each flac file)
}
\description{
Write effort pipeline csv files
}
