% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_birdnet_json_2.R
\name{read_birdnet_json_2}
\alias{read_birdnet_json_2}
\title{Title}
\usage{
read_birdnet_json_2(json_gz_path, species_threshold_df, p_true_positive)
}
\arguments{
\item{p_true_positive}{}
}
\description{
Title
}
