% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cb.R
\name{theme_cb}
\alias{theme_cb}
\title{CAbioacoustics plotting theme}
\usage{
theme_cb()
}
\value{
A \code{ggplot2} object with the \code{CAbioacoustics} plotting theme
}
\description{
CAbioacoustics plotting theme
}
\examples{
\dontrun{
# mtcars plot
library(ggplot2)

ggplot(mtcars, aes(x = hp, y = mpg, color = cyl, shape = as.factor(cyl))) +
  geom_point(size = 3) +
  theme_cb()
}
}
