% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_build_s3_db.R
\name{cb_build_s3_db}
\alias{cb_build_s3_db}
\title{Create a database of S3 contents that is easy to query}
\usage{
cb_build_s3_db(s3_df)
}
\description{
Create a database of S3 contents that is easy to query
}
