% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_check_empty_folders.R
\name{cb_check_empty_folders}
\alias{cb_check_empty_folders}
\title{Check for folders missing acoustic data on research drive}
\usage{
cb_check_empty_folders(path, year)
}
\arguments{
\item{year}{}
}
\description{
Check for folders missing acoustic data on research drive
}
