% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_selection_directories.R
\name{cb_create_selection_directories}
\alias{cb_create_selection_directories}
\title{Create directories to store selection tables}
\usage{
cb_create_selection_directories(date_time, directory)
}
\arguments{
\item{directory}{}
}
\description{
Create directories to store selection tables
}
