% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_selection_table.R
\name{cb_create_selection_table}
\alias{cb_create_selection_table}
\title{Title}
\usage{
cb_create_selection_table(df, date_time, species_group)
}
\arguments{
\item{df}{}
}
\description{
Title
}
