% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_swift_labels.R
\name{cb_create_swift_labels}
\alias{cb_create_swift_labels}
\title{Create PDF of scannable QR codes for Swift recorders (geared toward Avery 5520 labels)}
\usage{
cb_create_swift_labels(
  swift_df,
  type = "matrix",
  file,
  font_size,
  label_height_ratio
)
}
\arguments{
\item{label_height_ratio}{}
}
\description{
Create PDF of scannable QR codes for Swift recorders (geared toward Avery 5520 labels)
}
