% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_swift_labels.R
\name{cb_create_swift_labels}
\alias{cb_create_swift_labels}
\title{Create PDF of Swift barcodes or QR codes}
\usage{
cb_create_swift_labels(swift_df, type, file)
}
\arguments{
\item{file}{}
}
\description{
Create PDF of Swift barcodes or QR codes
}
