% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_delete_db_records.R
\name{cb_delete_db_records}
\alias{cb_delete_db_records}
\title{Delete a record from the database}
\usage{
cb_delete_db_records(ids, table_name)
}
\arguments{
\item{table_name}{}
}
\description{
Delete a record from the database
}
