% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_extract_json_predictions.R
\name{cb_extract_json_predictions}
\alias{cb_extract_json_predictions}
\title{Title}
\usage{
cb_extract_json_predictions(path, species_thresholds = species_threshold_df)
}
\arguments{
\item{species_thresholds}{}
}
\description{
Title
}
