% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_extract_json_predictions.R
\name{cb_extract_json_predictions}
\alias{cb_extract_json_predictions}
\title{Extract predictions from JSON files using species-specific threshold values}
\usage{
cb_extract_json_predictions(json, threshold_df, predictions_output, log_output)
}
\arguments{
\item{log_output}{}
}
\description{
Extract predictions from JSON files using species-specific threshold values
}
