% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_extract_raster.R
\name{cb_extract_raster}
\alias{cb_extract_raster}
\title{Title}
\usage{
cb_extract_raster(
  sf_object,
  scale,
  raster_layer,
  raster_name,
  buffer_size = NULL,
  fun = NULL
)
}
\arguments{
\item{buffer}{}
}
\description{
Title
}
