% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_flac_file_structure.R
\name{cb_flac_file_structure}
\alias{cb_flac_file_structure}
\title{Check that flacs are properly nested on research drive}
\usage{
cb_flac_file_structure(network_drive, survey_year, regions)
}
\arguments{
\item{regions}{}
}
\description{
Check that flacs are properly nested on research drive
}
