% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_flac_file_structure.R
\name{cb_flac_file_structure}
\alias{cb_flac_file_structure}
\title{Create expected FLAC file path based on drive location, survey year, and study regions}
\usage{
cb_flac_file_structure(network_drive, survey_year, regions)
}
\arguments{
\item{regions}{}
}
\description{
Create expected FLAC file path based on drive location, survey year, and study regions
}
