% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_audio_duration.R
\name{cb_get_audio_duration}
\alias{cb_get_audio_duration}
\title{Get duration of audio file (in seconds)}
\usage{
cb_get_audio_duration(path)
}
\arguments{
\item{path}{Path to audio file}
}
\value{
Duration of audio file in seconds
}
\description{
Get duration of audio file (in seconds)
}
