% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_flac_paths.R
\name{cb_get_flac_paths}
\alias{cb_get_flac_paths}
\title{Get path info for FLACs on research drive}
\usage{
cb_get_flac_paths(path, year)
}
\arguments{
\item{year}{}
}
\description{
Get path info for FLACs on research drive
}
