% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_hd_contents.R
\name{cb_get_hd_contents}
\alias{cb_get_hd_contents}
\title{Get FLAC information from a hard drive}
\usage{
cb_get_hd_contents(drive_path, output_dir)
}
\arguments{
\item{output_dir}{}
}
\description{
Get FLAC information from a hard drive
}
