% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_json_paths.R
\name{cb_get_json_paths}
\alias{cb_get_json_paths}
\title{Get path info for JSONs on research drive}
\usage{
cb_get_json_paths(path, year)
}
\arguments{
\item{path}{Path to JSON files}

\item{year}{Year}
}
\value{
Tibble of JSON file paths with size (MB), file name, and deployment name
}
\description{
Get path info for JSONs on research drive
}
