% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_globus_transfer.R
\name{cb_globus_transfer}
\alias{cb_globus_transfer}
\title{Execute a Globus file transfer between two collections from R}
\usage{
cb_globus_transfer(
  collection_one,
  collection_two,
  destination_path,
  batch_file_path,
  label
)
}
\arguments{
\item{label}{}
}
\description{
Execute a Globus file transfer between two collections from R
}
