% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_json_file_structure.R
\name{cb_json_file_structure}
\alias{cb_json_file_structure}
\title{Title}
\usage{
cb_json_file_structure(network_drive, survey_year, regions)
}
\arguments{
\item{regions}{}
}
\description{
Title
}
