% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_make_aru_sf.R
\name{cb_make_aru_sf}
\alias{cb_make_aru_sf}
\title{Make a data frame containing ARUs and UTMs into a simple feature object}
\usage{
cb_make_aru_sf(df)
}
\arguments{
\item{df}{}
}
\description{
Make a data frame containing ARUs and UTMs into a simple feature object
}
\examples{
\dontrun{
# example ARU data frame
aru_df <-
tibble::tribble(
  ~cell_id, ~utm_zone, ~utme, ~utmn,
  'C0001', 10, 670282, 4390557,
  'C0001', 10, 670972, 4390553,
  'C0001', 10, 669829, 4390388
)
# convert to simple feature point geometry
aru_df |>
dplyr::group_split(utm_zone) |>
  purrr::map_dfr(cb_make_aru_sf)
}
}
