% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_occ_start_end_dates.R
\name{cb_occ_start_end_dates}
\alias{cb_occ_start_end_dates}
\title{Get start and end dates and number of secondary sampling occasions based off annual monitoring and length of sampling occasions}
\usage{
cb_occ_start_end_dates(aru_dates_df, occasion_length)
}
\arguments{
\item{aru_dates_df}{A data frame of ...}

\item{occasion_length}{Length (in days) of secondary sampling occasions}
}
\value{
A data frame of start/end dates and the number of secondary sampling occasions
}
\description{
Get start and end dates and number of secondary sampling occasions based off annual monitoring and length of sampling occasions
}
