% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_read_field_maps.R
\name{cb_read_field_maps}
\alias{cb_read_field_maps}
\title{Read ARU deployment/recovery data from ArcGIS Field Maps}
\usage{
cb_read_field_maps(feature_layer)
}
\arguments{
\item{feature_layer}{HTML to hosted Field Maps feature layer}
}
\description{
Read ARU deployment/recovery data from ArcGIS Field Maps
}
