% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_read_filter_json.R
\name{cb_read_filter_json}
\alias{cb_read_filter_json}
\title{Title}
\usage{
cb_read_filter_json(path, species_thresholds = species_threshold_df)
}
\arguments{
\item{species_thresholds}{}
}
\description{
Title
}
