% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_rename_files.R
\name{cb_rename_flacs}
\alias{cb_rename_flacs}
\title{Rename FLAC files with incorrect deployment information}
\usage{
cb_rename_flacs(dir, pattern, replacement)
}
\arguments{
\item{dir}{Directory where FLACs are kept (for example, folder containing all FLACs for an ARU deployment)}

\item{pattern}{String needing replacing}

\item{replacement}{String to replace it with}
}
\value{
Renamed FLAC files in renamed directories
}
\description{
Rename FLAC files with incorrect deployment information
}
\examples{
\dontrun{
# correct G001_C0001_V1_U2 to G001_C0001_V1_U1
cb_rename_flacs(
  dir = 'path/to/deployments/G001_C0001_V1_U2',
  pattern = 'U2',
  replacement = 'U1'
  )
}
}
