% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_sampling_summary.R
\name{cb_sampling_summary}
\alias{cb_sampling_summary}
\title{Get summary of sampling effort}
\usage{
cb_sampling_summary(
  species,
  template,
  study_type,
  cell_ids,
  start_year,
  end_year
)
}
\arguments{
\item{end_year}{}
}
\description{
Get summary of sampling effort
}
