% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_segment_flac.R
\name{cb_segment_flac}
\alias{cb_segment_flac}
\title{Segment a flac file using the begin/end time of a BirdNET selection}
\usage{
cb_segment_flac(begin_file, begin_time, end_time)
}
\arguments{
\item{begin_file}{Name of flac file}

\item{begin_time}{Time from start of recording and beginning of a selection}

\item{end_time}{Time from start of recording and end of selection}
}
\description{
Segment a flac file using the begin/end time of a BirdNET selection
}
