% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_segment_flac.R
\name{cb_segment_flac}
\alias{cb_segment_flac}
\title{Segment a FLAC file using the begin/end time of a BirdNET selection}
\usage{
cb_segment_flac(input_flac_path, output_flac_path, begin_time, end_time)
}
\arguments{
\item{input_flac_path}{Path for input FLAC file}

\item{output_flac_path}{Path to store FLAC segment}

\item{begin_time}{Time from start of recording (beginning of a selection)}

\item{end_time}{Time from start of recording (end of selection)}
}
\value{
FLAC file segment
}
\description{
Segment a FLAC file using the begin/end time of a BirdNET selection
}
\examples{
\dontrun{
# set FLAC parameters
input_flac_path <- "Z:/Acoustic_Data/ARU_Data_Raw/ARU_Sierra_Monitoring/2025/South/G055_V1/G055_V1_C4928_U1/G055_V1_C4928_U1_20250519/G055_V1_C4928_U1_20250519_060002Z.flac"
output_flac_path <- "C:/Users/jmwin/OneDrive/Desktop/test_segment.flac"
begin_time <- 3
end_time <- 6

# create FLAC segment
cb_segment_flac(
  input_flac_path,
  output_flac_path,
  begin_time,
  end_time
)
}
}
