% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_set_db_credentials.R
\name{cb_set_db_credentials}
\alias{cb_set_db_credentials}
\title{Establish credentials for the UW spotted owl database}
\usage{
cb_set_db_credentials()
}
\value{
Saved passwords for spotted owl database
}
\description{
Establish credentials for the UW spotted owl database
}
\examples{
\dontrun{
# this will prompt to enter 1) the host, 2) user, 3) password, and 4) database name
cb_set_db_credentials()
}
}
