% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_set_future_cores.R
\name{cb_set_future_cores}
\alias{cb_set_future_cores}
\title{Set number of cores to run functions in parallel with future/furrr}
\usage{
cb_set_future_cores(n)
}
\arguments{
\item{n}{Number of cores to use. If blank, use all except leave one available for other tasks.}
}
\value{
Number of cores to use in parallel
}
\description{
Set number of cores to run functions in parallel with future/furrr
}
\examples{
\dontrun{
# use all cores - 1
cb_set_future_cores()

# or a specified number of cores
cb_set_future_cores(n = 10)
}
}
