% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_wav_to_flac.R
\name{cb_wav_to_flac}
\alias{cb_wav_to_flac}
\title{Convert WAV files stored on Swift SD card to FLAC files stored on external hard drive}
\usage{
cb_wav_to_flac(
  desktop_path,
  sd_card_path,
  hard_drive_path,
  sox_path,
  deployment_df
)
}
\arguments{
\item{deployment_df}{}
}
\description{
Convert WAV files stored on Swift SD card to FLAC files stored on external hard drive
}
